+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version 0.90-1.3.0
 
 * Fix a translation bug introduce in 0.90-1.2..5 that include the impossibility to save or update forms destinations.
 * Fix validation link in notifications (now set with Configuration value instead of fixed value)
+* Fix notification on ticket creation for destination with only one requester.
 
 ### Features: